Major changes to 'Home.vue' Moved items from Home into a new file 'components/Home/HomeButtonContainer.vue'
Temporarily removed the INSANE regex from 'UserProfile.vue',

Tried to fix 'Messages.vue'
