+++ java/database/create backup.sh
@@ -1,0 +1,9 @@
+#!/bin/bash
+export PGPASSWORD='postgres1'
+BASEDIR=$(dirname $0)
+DATABASE=final_capstone
+psql -U postgres -f "$BASEDIR/dropdb.sql" &&
+createdb -U postgres $DATABASE &&
+psql -U postgres -d $DATABASE -f "$BASEDIR/schema.sql" &&
+psql -U postgres -d $DATABASE -f "$BASEDIR/data.sql" &&
+psql -U postgres -d $DATABASE -f "$BASEDIR/user.sql"
