+++ java/src/main/resources/application.properties
@@ -1,4 +1,4 @@
-# datasource connection properties
+ # datasource connection properties
 spring.datasource.url=jdbc:postgresql://localhost:5432/final_capstone
 spring.datasource.name=final_capstone
 spring.datasource.username=final_capstone_appuser
