+++ vue/src/components/Home/HomeSlideshow.vue
